Restructure root navigator into stack -> tab format; Import main routes;
