+++ src/navigation/AppNavigator.js
@@ -1,102 +1,153 @@
 import React from 'react'
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs'
+import { createStackNavigator, HeaderStyleInterpolators } from '@react-navigation/stack'
 
-import FeedNavigator from 'navigation/Feed'
-import SearchNavigator from 'navigation/Search'
-import CameraNavigator from 'navigation/Camera'
-import DatingNavigator from 'navigation/Dating'
-import ProfileNavigator from 'navigation/Profile'
+import TabNavigator from 'navigation/TabNavigator'
 
-import HomeIcon from 'assets/svg/footer/Home'
-import SearchIcon from 'assets/svg/footer/Search'
-import CreateIcon from 'assets/svg/footer/Create'
-import HeartIcon from 'assets/svg/footer/Heart'
-import UserIcon from 'assets/svg/footer/User'
+import ModalScreen from 'screens/ModalScreen'
+import StoryScreen from 'screens/StoryScreen'
+import PostCreateScreen from 'screens/PostCreateScreen'
+import PostEditScreen from 'screens/PostEditScreen'
+import PostMediaScreen from 'screens/PostMediaScreen'
+import PostMediaLikesScreen from 'screens/PostMediaLikesScreen'
+import PostMediaViewsScreen from 'screens/PostMediaViewsScreen'
+import PostShareScreen from 'screens/PostShareScreen'
+import ProfileScreen from 'screens/ProfileScreen'
+import AlbumScreen from 'screens/AlbumScreen'
+import AlbumsScreen from 'screens/AlbumsScreen'
+import AlbumCreateScreen from 'screens/AlbumCreateScreen'
+import CommentsScreen from 'screens/CommentsScreen'
+import ProfileFollowedScreen from 'screens/ProfileFollowedScreen'
+import ProfileFollowerScreen from 'screens/ProfileFollowerScreen'
+import ProfileRequestsScreen from 'screens/ProfileRequestsScreen'
+import VerificationScreen from 'screens/VerificationScreen'
 
-
-const AppNavigator = ({ theme, ...props }) => {
-  const tabNavigatorProps = {
-    tabBarOptions: {
-      showLabel: false,
-      activeTintColor: theme.colors.primaryIcon,
-      style: {
-        backgroundColor: theme.colors.backgroundPrimary,
-        borderTopWidth: 0,
-      },
+const AppNavigator = ({ theme }) => {
+  const Stack = createStackNavigator()
+  const stackNavigatorProps = {
+    screenOptions: {
+      headerStyleInterpolator: HeaderStyleInterpolators.forSlideUp,
     },
   }
 
-  const FeedTabIconComponent = ({ color }) => <HomeIcon fill={color} />
-  const feedTabScreenPropsCard = {
+  const stackScreenProps = {
     options: {
-      tabBarIcon: FeedTabIconComponent,
-    },
-  }
-  
-  const SearchTabIconComponent = ({ color }) => <SearchIcon fill={color} />
-  const searchTabScreenPropsCard = {
-    options: {
-      tabBarIcon: SearchTabIconComponent,
+      headerShown: false,
     },
   }
 
-  const CameraTabIconComponent = ({ color }) => <CreateIcon fill={color} />
-  const cameraTabScreenPropsCard = {
-    options: {
-      tabBarIcon: CameraTabIconComponent,
-    },
-  }
+  const Tab = TabNavigator(theme)
 
-  const DatingTabIconComponent = ({ color }) => <HeartIcon fill={color} />
-  const datingTabScreenPropsCard = {
-    options: {
-      tabBarIcon: DatingTabIconComponent,
-    },
-  }
-
-  const ProfileTabIconComponent = ({ color }) => <UserIcon fill={color} />
-  const profileTabScreenPropsCard = {
-    options: {
-      tabBarIcon: ProfileTabIconComponent,
-    },
-  }
-
-  const Tab = createBottomTabNavigator()
-  
-  const Feed = FeedNavigator(theme)
-  const Search = SearchNavigator(theme)
-  const Camera = CameraNavigator(theme)
-  const Dating = DatingNavigator(theme)
-  const Profile = ProfileNavigator(theme)
-  
   return (
-    <Tab.Navigator {...tabNavigatorProps}>
-      <Tab.Screen
-        name="Feed"
-        component={Feed}
-        {...feedTabScreenPropsCard}
+    <Stack.Navigator {...stackNavigatorProps}>
+      <Stack.Screen
+        name="Home"
+        component={Tab}
+        {...stackScreenProps}
       />
-      <Tab.Screen
-        name="Search"
-        component={Search}
-        {...searchTabScreenPropsCard}
+
+      <Stack.Screen
+        name="Modal"
+        component={ModalScreen}
+        {...stackScreenProps}
       />
-      <Tab.Screen
-        name="Camera"
-        component={Camera}
-        {...cameraTabScreenPropsCard}
+
+      <Stack.Screen
+        name="Story"
+        component={StoryScreen}
+        {...stackScreenProps}
       />
-      <Tab.Screen
-        name="Dating"
-        component={Dating}
-        {...datingTabScreenPropsCard}
+
+      <Stack.Screen
+        name="PostCreate"
+        component={PostCreateScreen}
+        {...stackScreenProps}
       />
-      <Tab.Screen
+
+      <Stack.Screen
+        name="PostEdit"
+        component={PostEditScreen}
+        {...stackScreenProps}
+      />
+
+      <Stack.Screen
+        name="PostMedia"
+        component={PostMediaScreen}
+        {...stackScreenProps}
+      />
+
+      <Stack.Screen
+        name="PostMediaLikes"
+        component={PostMediaLikesScreen}
+        {...stackScreenProps}
+      />
+
+      <Stack.Screen
+        name="PostMediaViews"
+        component={PostMediaViewsScreen}
+        {...stackScreenProps}
+      />
+
+      <Stack.Screen
+        name="PostShare"
+        component={PostShareScreen}
+        {...stackScreenProps}
+      />
+
+      <Stack.Screen
         name="Profile"
-        component={Profile}
-        {...profileTabScreenPropsCard}
+        component={ProfileScreen}
+        {...stackScreenProps}
       />
-    </Tab.Navigator>
+
+      <Stack.Screen
+        name="Album"
+        component={AlbumScreen}
+        {...stackScreenProps}
+      />
+
+      <Stack.Screen
+        name="Albums"
+        component={AlbumsScreen}
+        {...stackScreenProps}
+      />
+
+      <Stack.Screen
+        name="AlbumCreate"
+        component={AlbumCreateScreen}
+        {...stackScreenProps}
+      />
+
+      <Stack.Screen
+        name="Comments"
+        component={CommentsScreen}
+        {...stackScreenProps}
+      />
+
+      <Stack.Screen
+        name="ProfileFollowed"
+        component={ProfileFollowedScreen}
+        {...stackScreenProps}
+      />
+
+      <Stack.Screen
+        name="ProfileFollower"
+        component={ProfileFollowerScreen}
+        {...stackScreenProps}
+      />
+
+      <Stack.Screen
+        name="ProfileRequests"
+        component={ProfileRequestsScreen}
+        {...stackScreenProps}
+      />
+
+      <Stack.Screen
+        name="Verification"
+        component={VerificationScreen}
+        {...stackScreenProps}
+      />
+
+    </Stack.Navigator>
   )
 }
