--- src/navigation/Navigator.js
@@ -1,1 +1,0 @@
-import { createStackNavigator } from '@react-navigation/stack'
